Reorganize file directory, add builder configuration, add introduction and simple display
