+++ README.md
@@ -1,51 +1,15 @@
-# 语料库
+# 《醒世姻缘转》
+方言词及熟语方言性研究小组
 
-# usage:
-```shell
-npm i  / cnpm i 
-npm run watch    //webpack监听打包（运行后另起一个shell运行下面的）
-./node_modules/.bin/elctron_rebuilt  //可能会有本地node-gyp构建工具依赖问题（大概率不会）
-npm start
-```
+## 简介
+《醒世姻缘传》作为十七世纪的一部重要世情小说，于明末清初用山东地区方言写成，书中保留大量口语词汇。在汉语史中口语和文言长期并存的状况下，《醒世姻缘传》使用并反映该时期及地域的口语，因此能够较为真实地反映断代方言词汇使用面貌。本项目以《醒世姻缘传》一书为语料来源，对其中的方言词进行词汇层面研究，对其中熟语的方言特征进行考察与研究，并建立专书语料库。
 
+## 展示
+### 正常阅读《醒世姻缘转》
+![](assets/show_assets/gif/reading.gif)
+### 搜索你想知道的词意
+![](assets/show_assets/gif/search.gif)
+### 跳转到文本部分，进行阅读
+![](assets/show_assets/gif/jump.gif)
 
-# 语料库实现路径
-
-我们选择使用 `Electron` 和 `React` 开发桌面应用程序，因为这两个技术都是目前市场上非常流行且功能强大的。Electron 可以让我们轻松地将应用程序部署到多个平台，如 Windows、macOS 和 Linux，而 React 则能让我们轻松地构建出易于维护和扩展的用户界面。同时，这两个技术的社区非常活跃，有大量的资源和文档可供参考，以便我们在遇到问题时能够快速找到解决方案。
-
-在应用程序的界面设计方面，我们选择使用了 `Material-UI`，它是一个基于 React 的流行 UI 框架。Material-UI 提供了许多现成的组件和样式，可以帮助我们快速地构建出美观且响应迅速的界面，提高用户体验。
-
-为了实现在应用程序中阅读书籍的功能，我们选用了react-pdf-viewer 和 pdfjs-dist 这两个库。这些库可以让我们方便地将书籍以 PDF 格式嵌入到应用程序中，并提供了丰富的阅读功能，如缩放、搜索和导航等。此外，我们还可以利用这些库提供的接口实现定位功能，使用户能够快速跳转到书籍中的相关内容。
-
-对于词汇检索和释义查询功能，我们选择使用了 better-sqlite3 这个数据库库。这个库提供了高性能的 SQLite 数据库支持，使我们能够方便地存储、检索和管理词汇及其释义。此外，SQLite 是一个轻量级的数据库，它可以直接嵌入到应用程序中，不需要额外的服务器或设置，这将大大简化部署和维护过程。
-
-**实现路径大致如下**：
-
-1. 设计和构建应用程序的基本结构，包括主界面、菜单和其他基本组件。
-2. 集成 react-pdf-viewer 和 pdfjs-dist，实现在应用程序中阅读书籍的功能。
-3. 使用 Material-UI 构建输入框和结果列表等界面组件，实现词汇查询功能。
-4. 集成 better-sqlite3，构建一个包含词汇和释义的数据库，并实现对数据库的查询功能。这将使得用户可以通过输入词汇快速查找到相应的释义。
-5. 通过利用 react-pdf-viewer 提供的接口，实现从查询结果直接跳转到书籍中相应的位置。当用户点击某个查询结果时，应用程序将自动导航到书籍中该词汇出现的位置，提高用户的查找效率。
-6. 对整个应用程序进行测试，确保所有功能都能正常运行，同时优化性能和用户体验。
-7. 完成开发后，我们将使用 Electron 打包应用程序，以便可以在 Windows、macOS 和 Linux 平台上运行。
-
-***具体技术细节***：
-
-1. **项目脚手架和环境搭建**：使用 **`create-react-app`** 或手动配置 Webpack、Babel、ESLint 等工具来搭建项目基础结构。在 **`.babelrc`** 中配置 **`@babel/preset-env`** 和 **`@babel/preset-react`**，以支持最新的 ECMAScript 和 React JSX 语法。
-2. **构建应用程序主体结构**：实现应用程序的主要布局，使用 React Hook（如 **`useState`** 和 **`useEffect`**）管理应用程序的状态。利用 React Context API 和 **`useContext`** Hook 在组件之间共享全局数据。
-3. **实现虚拟 DOM 渲染优化**：利用 React 的 **`memo`** 和 **`useCallback`** Hook 函数，避免不必要的重新渲染和函数重建，从而提高应用程序的性能。
-4. **集成 Material-UI**：在项目中引入 **`@material-ui/core`** 和 **`@material-ui/icons`**，使用 **`makeStyles`** 和 **`ThemeProvider`** 进行自定义样式和主题设置。
-5. **实现 PDF 阅读功能**：在 React 组件中引入 **`@react-pdf-viewer/core`** 和 **`@react-pdf-viewer/default-layout`**，配置 PDF.js worker 文件，使用 **`Document`** 和 **`Page`** 组件进行 PDF 文件的渲染和翻页操作。
-6. **数据库设计与集成**：创建一个 SQLite 数据库，设计合适的表结构（例如，使用外键约束和多对多关系），并为词汇及其释义、出现位置等信息建立索引以提高查询性能。在项目中引入 **`better-sqlite3-with-prebuilds`**，实现数据的 CRUD（创建、读取、更新和删除）操作。
-7. **词汇检索组件实现**：使用 React 的受控组件模式和防抖（debouncing）技术来实现实时搜索功能。为词汇条目添加 **`React.Fragment`** 和 **`key`** 属性，以优化渲染性能。
-8. **利用虚拟列表优化长列表性能**：对于搜索结果的长列表，使用 **`react-window`** 或 **`react-virtualized`** 实现虚拟列表，以减少 DOM 节点的数量并提高渲染性能。
-9. **使用 React Portal 实现弹窗**：利用 React Portal 技术在 DOM 层次结构之外渲染词汇释义弹窗，以避免样式冲突和提高组件独立性。
-10. **实现词汇高亮与定位**：通过扩展 **`react-pdf-viewer`** 的自定义渲染层，实现在 PDF 文档中对词汇的高亮显示。在数据库查询结果中返回词汇在文档中的位置信息（如页码和坐标），并利用这些信息进行定位和跳转。结合 Intersection Observer API，实现在滚动阅读时动态高亮当前显示的词汇。
-11. **代码分割与懒加载**：使用 Webpack 的代码分割功能，将应用程序划分为多个较小的代码块。结合 React 的 **`lazy()`** 和 **`Suspense`** API，实现组件和库的按需加载，以减少首次加载时间。
-12. **错误处理与异常捕获**：使用 React 的错误边界（Error Boundary）机制来捕获和处理组件树中的错误。对于异步操作（如数据库查询和 PDF 渲染），使用 **`try-catch`** 语句或 Promise 的 **`catch()`** 方法进行错误处理。
-13. **性能监控与优化**：使用 Chrome DevTools 或其他性能分析工具，分析应用程序的运行时性能，包括内存使用、渲染时间和资源加载等。根据分析结果进行性能优化，例如减少不必要的组件渲染、合并 CSS 样式、压缩静态资源等。
-14. **版本控制和管理**：撰写项目文档，并结合 GitLab、GitHub 等版本控制平台进行管理和追踪。
-15. **应用程序打包与发布**：利用 Electron 的 **`electron-builder`** 或 **`electron-packager`** 等工具，将应用程序打包为可执行文件（如 **`.exe`**、**`.dmg`** 和 **`.deb`** 等格式）。在 Windows、macOS 和 Linux 等目标平台上进行测试，确保应用程序的兼容性和稳定性。
-
-总之，我们的技术选型和实现路径将充分利用现有的优秀库和框架，以简化开发过程并确保应用程序的高质量。将这些库和框架结合到一起以构建一个功能完善、易于使用的应用程序过程中我们解决了诸多冲突依赖等问题。我相信，我们的努力将提供一个优秀的语料库程序。
-
+## 实现细节
