--- index.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <meta charset="UTF-8">
-  <!-- https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP -->
-  <title>语料库</title>
-</head>
-
-<body>
-  <div id="app"></div>
-  <script src="./app.js"></script>
-  <div id="trans">
-
-  </div>
-  <script src="./renderer.js"></script>
-</body>
-
-</html>
