--- main.js
@@ -1,57 +1,0 @@
-// main.js
-
-// Modules to control application life and create native browser window
-const { app, BrowserWindow } = require('electron')
-//const path = require('path')
-
-
-
-const createWindow = () => {
-    if(process.platform == 'darwin' || process.platform == "linux"){
-        preload_path = `${__dirname}/preload.js`
-    }else if(process.platform == 'win32'){
-        preload_path = `${__dirname}\\preload.js`
-    }
-  // Create the browser window.
-  const mainWindow = new BrowserWindow({
-    width: 800,
-    height: 600,
-    webPreferences: {
-      nodeIntegration:true,
-      //preload: path.join(__dirname, './preload.js')
-      preload: preload_path
-    }
-  })
-  //mainWindow.webContents.openDevTools()
-  // and load the index.html of the app.
-  mainWindow.loadFile('index.html')
-
-  // Open the DevTools.
-  // mainWindow.webContents.openDevTools()
-}
-
-// This method will be called when Electron has finished
-// initialization and is ready to create browser windows.
-// Some APIs can only be used after this event occurs.
-app.whenReady().then(() => {
-  createWindow()
-
-  app.on('activate', () => {
-    // On macOS it's common to re-create a window in the app when the
-    // dock icon is clicked and there are no other windows open.
-    if (BrowserWindow.getAllWindows().length === 0) createWindow()
-  })
-})
-
-// Quit when all windows are closed, except on macOS. There, it's common
-// for applications and their menu bar to stay active until the user quits
-// explicitly with Cmd + Q.
-app.on('window-all-closed', () => {
-  if (process.platform !== 'darwin') app.quit()
-})
-
-
-
-
-// In this file you can include the rest of your app's specific main process
-// code. You can also put them in separate files and require them here.
