+++ package.json
@@ -1,63 +1,45 @@
 {
-  "name": "corpus",
-  "version": "1.0.0",
-  "description": "An Corpus",
-  "main": "main.js",
-  "scripts": {
-    "start": "electron .",
-    "watch": "webpack --config webpack.common.js --watch",
-    "postinstall": "electron-builder install-app-deps"
-  },
-  "biuld": {
-    "appId": "com.electron.Corpus",
-    "productName": "Corpus",
-    "copyright": "Copyright © 2023 corpus",
-    "nodeGypRebuild": true,
-    "win": {
-      "target": [
-        "portable"
-      ],
-      "icon": "build/icon.png"
+    "name": "corpus",
+    "version": "1.0.0",
+    "description": "An Corpus",
+    "main": "./src/main/main.js",
+    "scripts": {
+        "start": "electron .",
+        "watch": "webpack --config ./config/webpack.common.js --watch",
+        "postinstall": "electron-builder install-app-deps",
+        "build": "electron-builder -c ./config/ele-builder.json"
     },
-    "mac": {
-      "target": [
-        "dmg",
-        "zip"
-      ],
-      "icon": "build/icon.png"
+    "license": "ISC",
+    "devDependencies": {
+        "@babel/core": "^7.21.3",
+        "@babel/preset-env": "^7.20.2",
+        "@babel/preset-react": "^7.18.6",
+        "babel-loader": "^9.1.2",
+        "css-loader": "^6.7.3",
+        "electron": "^16.2.8",
+        "electron-builder": "^23.6.0",
+        "electron-rebuild": "^3.2.9",
+        "sass": "^1.59.3",
+        "sass-loader": "^13.2.0",
+        "style-loader": "^3.3.2",
+        "webpack": "^5.76.2",
+        "webpack-cli": "^5.0.1"
+    },
+    "dependencies": {
+        "@electron/asar": "^3.2.3",
+        "@material-ui/core": "^4.12.4",
+        "@material-ui/icons": "^4.11.3",
+        "@react-pdf-viewer/core": "^3.11.0",
+        "@react-pdf-viewer/default-layout": "^3.11.0",
+        "@react-pdf-viewer/search": "^3.11.0",
+        "better-sqlite3": "^8.2.0",
+        "pdfjs-dist": "3.3.122",
+        "react": "^17.0.0",
+        "react-dom": "^17.0.0",
+        "react-modal": "^3.16.1",
+        "react-router-dom": "^6.9.0",
+        "react-scrollbars-custom": "^4.1.1",
+        "react-spinners": "^0.13.8",
+        "react-spring": "^9.7.1"
     }
-  },
-  "license": "ISC",
-  "devDependencies": {
-    "@babel/core": "^7.21.3",
-    "@babel/preset-env": "^7.20.2",
-    "@babel/preset-react": "^7.18.6",
-    "babel-loader": "^9.1.2",
-    "css-loader": "^6.7.3",
-    "electron": "^16.2.8",
-    "electron-builder": "^23.6.0",
-    "electron-rebuild": "^3.2.9",
-    "sass": "^1.59.3",
-    "sass-loader": "^13.2.0",
-    "style-loader": "^3.3.2",
-    "webpack": "^5.76.2",
-    "webpack-cli": "^5.0.1"
-  },
-  "dependencies": {
-    "@electron/asar": "^3.2.3",
-    "@material-ui/core": "^4.12.4",
-    "@material-ui/icons": "^4.11.3",
-    "@react-pdf-viewer/core": "^3.11.0",
-    "@react-pdf-viewer/default-layout": "^3.11.0",
-    "@react-pdf-viewer/search": "^3.11.0",
-    "better-sqlite3": "^8.2.0",
-    "pdfjs-dist": "3.3.122",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
-    "react-modal": "^3.16.1",
-    "react-router-dom": "^6.9.0",
-    "react-scrollbars-custom": "^4.1.1",
-    "react-spinners": "^0.13.8",
-    "react-spring": "^9.7.1"
-  }
 }
