--- renderer.js
@@ -1,12 +1,0 @@
-// document.addEventListener('DOMContentLoaded',async()=>{
-//     let names = window.api.getTrans();
-//     console.log(names);
-//     let divNames = document.getElementById('trans');
-
-//     let nameString = names.map((elem)=>{
-//         return elem.key;
-//     }).join("<br />");
-
-//     // let nameString = names.join("<br />");
-//     divNames.innerHTML = nameString;
-// })
