--- src/index.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import MainPage from './MainPage';
-window.onload = () => {
-    ReactDOM.render(<MainPage/>, document.getElementById('app'));
-};
