+++ src/main/index.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+  <meta charset="UTF-8">
+  <!-- https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP -->
+  <title>语料库</title>
+</head>
+
+<body>
+  <div id="app"></div>
+  <script src="./app.js"></script>
+  <div id="trans">
+
+  </div>
+</body>
+
+</html>
