+++ src/public/index.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import MainPage from './MainPage';
+window.onload = () => {
+    ReactDOM.render(<MainPage/>, document.getElementById('app'));
+};
